# feature(progress): re-vamps the performance log for ease of use