## Description

- adds _user_ and _system_ CPU timings
- changes display of memory values from _absolute_ to _relative_
- uses the `Intl` API to format the numbers - so it's closer to what the customer is used to
- adds the _total heap used_ to the summary on the last line

## Motivation and Context

- The _users_ and _system_ timings give a more fine grained view on what's happening under dependency-cruiser's hood
- When the log still had absolute values for memory usage I discovered I was continuously doing the math to check what the difference was in each step. 
- The `Intl` API also does away with some custom formatting that was still going on in the performance log helpers. Added bonus is that it takes the customer's locale into account making for much more recognizable (and readable) numbers.
- With the delta's the numbers displayed will be smaller as well, allowing to show memory sizes in kB instead of Mb - which gives a clearer view of what is actually going on.

## How Has This Been Tested?

- [x] green ci
- [x] adapted unit tests
- [x] manual usage (both on dependency-cruiser itself and on the facebook/react codebase)

## Screenshots

### After

Note: if the terminal supports color, the logger uses `chalk` to make things more readable a.o. coloring negative numbers and making titles bold.

```
$ node ./bin/dependency-cruise.js src bin test configs types tools --config --ignore-known
elapsed real         user       system        ∆ rss  ∆ heapTotal   ∆ heapUsed   ∆ external after step...
------------ ------------ ------------ ------------ ------------ ------------ ------------ ------------ 
       832ms        835ms        101ms   +118,396kB    +84,364kB    +59,833kB     +2,231kB start of node process
        11ms         10ms          1ms       +864kB       +256kB       +420kB          0kB parsing options
        91ms         22ms         11ms     +2,656kB       +544kB     +1,808kB       +583kB cache: checking freshness
...
```

### Before

```
$ node ./bin/dependency-cruise.js src bin test configs types tools --config --ignore-known
  elapsed       rss heapTotal  heapUsed  external after step...
    751ms     127Mb      84Mb      61Mb       2Mb start of node process
     11ms     128Mb      84Mb      62Mb       2Mb parsing options
     93ms     130Mb      85Mb      63Mb       2Mb cache: checking freshness
...
```

## Types of changes

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] Documentation only change
- [ ] Refactor (non-breaking change which fixes an issue without changing functionality)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)

## Checklist

- [x] :book:

  - My change doesn't require a documentation update, or ...
  - it _does_ and I have updated it

- [x] :balance_scale:
  - The contribution will be subject to [The MIT license](https://github.com/sverweij/dependency-cruiser/blob/develop/LICENSE), and I'm OK with that.
  - The contribution is my own original work.
  - I am ok with the stuff in [**CONTRIBUTING.md**](https://github.com/sverweij/dependency-cruiser/blob/develop/.github/CONTRIBUTING.md).